terminei de implementar o inventário e o combate e iniciei o ciclo dia noite e a tela de game over

+++ src/game/itens/weapons/Weapon.java
@@ -7,7 +7,7 @@ public abstract class Weapon extends Item {
     private int damage;
     private int range;
 
-    public Weapon(String name, int weight, int durability, String type, int damage, int range) {
+    public Weapon(String name, float weight, int durability, String type, int damage, int range) {
         super("weapon", name, weight, durability, true);
         this.type = type;
         this.damage = damage;
